refactor: remove unnecessary code, reorganise components (#114)
