--- src/components/Blogposts/blogposts.module.scss
@@ -1,22 +1,0 @@
-@import "../../styles/utility/responds-to.scss";
-
-.blogposts {
-  h2 {
-    margin-bottom: var(--title-spacing);
-  }
-  img {
-    margin-bottom: 0;
-  }
-}
-
-.flex {
-  @include responds-to("sm", "min") {
-    padding-right: 2rem;
-  }
-  @include responds-to("md", "min") {
-    padding-right: 6rem;
-  }
-  flex: 1 1 400px;
-  display: flex;
-  flex-direction: column;
-}
