+++ src/components/Header/index.js
@@ -1,20 +1,50 @@
+import { Link } from "gatsby"
+import PropTypes from "prop-types"
 import React from "react"
-import Center from "../Center"
+import MobileMenu from "../MobileMenu"
 
+// icons
+import logo from "../../icons/logo.svg"
 import styles from "./header.module.scss"
 
-const Index = () => (
-  <Center maxWidth="var(--content-width)" className={styles.header}>
-    <h1 className="xx-large-text">FreeCodeCamp Manchester</h1>
-    <p>
-      Welcome to FreeCodeCamp Manchester! We are mentors, coders and learners
-      from Manchester, Greater Manchester and its surrounding areas and one part
-      of the global FreeCodeCamp Community. We hold Study Groups twice a month
-      for our members to meet in person, code together and get help and advice
-      from Organisers and Mentors - Keep an eye on the events for further
-      information and we hope to see you around!
-    </p>
-  </Center>
+const activeStyles = {
+  textDecoration: "underline",
+}
+
+const NavLink = ({ children, to }) => (
+  <Link to={to} activeStyle={activeStyles}>
+    {children}
+  </Link>
 )
 
-export default Index
+const MainNav = () => (
+  <nav className={styles.nav}>
+    <NavLink to="/">Home</NavLink>
+    <NavLink to="/blog/">Blog</NavLink>
+    <NavLink to="/learning-resources/">Resources</NavLink>
+    <NavLink to="/upcoming-meetups/">Upcoming Meetups</NavLink>
+    <MobileMenu />
+  </nav>
+)
+
+const Header = ({ siteTitle }) => (
+  <>
+    <header className={`${styles.header}  u-shadow--1`}>
+      <Link to="/" className={styles.title} aria-label="Home">
+        {/* need to bring in icon */}
+        <img className={styles.logo} src={logo} alt="" />
+      </Link>
+      <MainNav />
+    </header>
+  </>
+)
+
+Header.propTypes = {
+  siteTitle: PropTypes.string,
+}
+
+Header.defaultProps = {
+  siteTitle: ``,
+}
+
+export default Header
