+++ src/components/Hero/hero.module.scss
@@ -1,0 +1,11 @@
+.hero {
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  margin-top: -2.499rem;
+  padding-top: 130px;
+
+  h1 {
+    margin-bottom: 1.2rem;
+  }
+}
